Split Types.hs into a few separate files
